feat: create TaskList class
- create taskList class to handle logic related to managing tasks
